Refactoring the tests models
