+++ qipr_approver/approver/tests/tests_models/test_project.py
@@ -8,7 +8,7 @@ from django.contrib.auth.models import User
 from django.core.urlresolvers import reverse
 from django.db.models import fields, signals
 from django.test import TestCase, Client
-from approver.utils import check_fields as check_fields
+from approver.tests.tests_models.test_utils import check_fields
 
 class ProjectTestCase(TestCase):
     def setUp(self):
@@ -16,5 +16,5 @@ class ProjectTestCase(TestCase):
         pass
 
     def test_project_model(self):
-        self.assertEqual(check_fields(Project,"title","Char",300), True)
-        self.assertEqual(check_fields(Project,"description","Text"), True)
+        self.assertTrue(check_fields(Project,"title","Char",300))
+        self.assertTrue(check_fields(Project,"description","Text"))
